Create own file for ResultWindowLimitError class
