+++ graylog2-web-interface/src/views/logic/SearchResult.js
@@ -1,7 +1,8 @@
 import { fromJS } from 'immutable';
 import { mapValues, get, compact } from 'lodash';
 import QueryResult from './QueryResult';
-import SearchError, { ResultWindowLimitError } from './SearchError';
+import SearchError from './SearchError';
+import ResultWindowLimitError from './ResultWindowLimitError';
 
 class SearchResult {
   constructor(result) {
